# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MagicAI  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documents/recent": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get the last 10 recent documents",
                "description": "Get the last 10 recent documents for the authenticated user, excluding documents of type 'image'.",
                "operationId": "getRecentDocs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/documents/all-openai": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Gets all AI Generators",
                "description": "Gets all AI Generators",
                "operationId": "getOpenAIList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/documents/openai-filters": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Gets all AI Filters",
                "description": "Gets all AI Filters",
                "operationId": "getOpenAIFilters",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/documents/": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Gets documents",
                "description": "Gets documents for the authenticated user, excluding documents of type 'image'.",
                "operationId": "getDocs",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter string defined in the 'name' column of the 'openai_filters' table.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort string. Values: 'newest', 'oldest', 'az', 'za'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "newest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/documents/doc/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Gets single document",
                "description": "Gets document by id, excluding documents of type 'image'.",
                "operationId": "getDoc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to get.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    },
                    "404": {
                        "description": "Document Not Found"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Updates single document",
                "description": "Updates document by id, excluding documents of type 'image'.",
                "operationId": "saveDoc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to get.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "output",
                        "in": "query",
                        "description": "Html Output of the document.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the document.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    },
                    "404": {
                        "description": "Document Not Found"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Deletes single document",
                "description": "Deletes document by id, excluding documents of type 'image'.",
                "operationId": "deleteDoc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to delete.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    },
                    "404": {
                        "description": "Document Not Found"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}